feat(app): Add "lakharg" Django app for searching and representing data
This commit adds a new Django app called "lakharg" to the project. The "lakharg" app is responsible for searching and representing data. It provides a set of tools to search and filter data from the database and represent it in different formats such as tables, charts, and graphs.

The addition of the "lakharg" app provides developers with a convenient way to search and represent data in the Django project.