refactor(backend): Remove unused files from the "lakhard" package
This commit refactors the "lakhard" app in the backend of the project by removing unused files. The removal of these files reduces clutter and helps to simplify the codebase.

The unused files include some test files and templates that are not being used in the app. These files can be safely removed without affecting the functionality of the app.