feat(backend): Add "lakhard" app to INSTALLED_APPS
This commit adds the "lakhard" app to the INSTALLED_APPS list in the Django settings module. The INSTALLED_APPS list is a list of strings representing the names of all Django applications that are installed in the project.

By adding "lakhard" to INSTALLED_APPS, Django will be able to recognize and use the app in the project. This is necessary for the app to be able to function properly and for its models and views to be accessible.