feat(backend): Create daily-updated words model for unique words
This commit adds a new words model to the backend of the project. The words model is responsible for maintaining a list of all unique words found in the text data stored in the backend. The model is updated daily to ensure that it always contains the most up-to-date list of words.

The words model is designed to be used in conjunction with the search API endpoint. By maintaining a list of all unique words, the search API endpoint can quickly and efficiently search through the text data to find relevant sentences.

The implementation of the words model involves parsing the text data stored in the backend and extracting all unique words. The words are then stored in the words model as a list. The words model is updated daily to ensure that it always contains the most up-to-date list of words.