feat(backend): Add endpoint to return statistics about unique words and books
This commit adds a new API endpoint to the backend of the project. The new endpoint is responsible for returning statistics about the number of unique words and books in the system.

When the endpoint is called, it performs a database query to count the number of unique words and books in the system. The endpoint then returns the statistics as a JSON object.

The addition of this endpoint provides a way for clients to retrieve information about the number of unique words and books in the system.