feat(backend): Add new model "Source" for saving information about sources
This commit adds a new model called "Source" to the backend of the project. The new model is responsible for storing information about sources of books, newspapers, and other types of publications.

The "Source" model includes fields for storing information such as the title of the publication, the name of the author, the date of publication, and the type of publication. The model also includes a URL field, which can be used to link to the source online.

The addition of the "Source" model provides a way to store and manage information about sources of publications in the system. This information can be useful for tracking the origin of particular pieces of content, as well as for providing attribution for quotes and other content that is used in the system.