chore: Add nltk and python-dotenv to requirements file
This commit adds two new dependencies, nltk and python-dotenv, to the requirements file for the project. The nltk library provides natural language processing tools that can be used for tasks such as tokenizing text, identifying parts of speech, and stemming words. The python-dotenv library provides a way to load environment variables from a .env file.

By adding these dependencies to the requirements file, we ensure that they will be installed when the project is set up or when the requirements are updated. This change is classified as a chore, as it is a maintenance task that does not introduce new functionality to the project.