feat: Add drf-spectacular and configuration
This commit adds the drf-spectacular package to the project and configures it to generate an OpenAPI schema for the API. drf-spectacular is a package that generates an OpenAPI schema for a Django REST Framework API, which can be used to generate API documentation and to test the API.

The configuration for drf-spectacular is added to the Django settings file, including the settings to enable schema generation and to set the location of the schema file. The schema file is configured to be generated at the '/api/schema/' endpoint.