fix: Fix path of favicon in index.html
This commit fixes the path of the favicon in the index.html file of the frontend application. Previously, the path was incorrect, which resulted in the favicon not being displayed correctly in the browser.

The new path is set to the correct location of the favicon file, which ensures that it will be displayed correctly in the browser.