feat: Add counter of available unique words to hero section of landing page
This commit adds a new feature to the hero section of the landing page. The feature is a counter that displays the current number of unique words that are available in the application. This provides users with a quick way to see how much data is available in the application.

The counter is implemented using React hooks and is displayed in the hero section of the landing page. It is updated automatically whenever new data is added to the application. The counter is also styled to fit in with the overall design of the landing page.