feat(frontend): add scrolling to tabs on mobile device
This commit adds the functionality to allow for scrolling between tabs on mobile devices. Previously, when there were more tabs than could fit on the screen, they were not accessible on mobile devices. This commit adds the necessary CSS and JavaScript to allow for scrolling between tabs on mobile devices.