feat(frontend): unmount inactive tabs to increase performance of the unique words tab
This commit improves the performance of the unique words tab by unmounting inactive tabs. This change is done to reduce the number of rendered components and increase the overall performance of the page. By unmounting the inactive tabs, we can free up the memory and resources that would have been used to render them, and allocate them to other components that are currently active.