feat(authguard): fix auth guard
Added a fix for the auth guard, which was not working properly due to some issues with the authentication provider. The fix involves updating the authentication guard logic to use a new method called isAuthenticated. This method is implemented in the new authentication provider and is used to check whether a user is authenticated or not. The fix also involves updating some of the existing code to use the new authentication guard logic.