feat(api): restrict API access to read-only
This commit adds functionality to restrict API access to read-only by modifying the permissions for the API endpoint. Now, only GET requests will be allowed to access the API, while all other HTTP methods will be disallowed. This feature enhances the security of the API by preventing unauthorized changes to the data.