chore(deps): upgrade mantine and mantine-datatable packages to latest version
The mantine and mantine-datatable packages have been upgraded to the latest version, which includes a lot of breaking changes that have been fixed. This upgrade ensures that the project is using the most up-to-date and stable versions of these packages, and helps to maintain compatibility with other dependencies in the project.