feat(configuration): add js-cookie package and configure axios to use CSRF token header
This commit adds the js-cookie package to the project's package.json file and configures Axios to use the CSRF token header. The js-cookie package is used to manage cookies in the application, while the CSRF token header is used to protect against cross-site request forgery (CSRF) attacks. This change improves the security of the application by ensuring that all requests made by Axios include the CSRF token header, making it more difficult for attackers to exploit vulnerabilities in the application.