fix(attribute): fix improperly used attribute after refactoring "publishers"
This commit fixes a bug where an attribute was improperly used after the "publishers" attribute was refactored. The original attribute was not updated to reflect the changes made to the "publishers" attribute, resulting in errors when the application attempted to use it. With this fix, the attribute is correctly updated and the application is able to use the new "publishers" attribute without issue.