fix(component): add missing JSX wrapper around Paper child elements
This commit fixes a bug in a component where a JSX wrapper was missing around Paper child elements. This was causing errors in the application and preventing the proper rendering of the component. With this fix, the missing JSX wrapper is added and the component is able to render properly as intended.