feat(database): display the actual number of diverse materials in the database
This commit adds functionality to the application to display the actual number of diverse materials in the database. Previously, the displayed number was not accurate and did not reflect the actual number of diverse materials in the database. With this new functionality, the application is able to provide users with an accurate count of the diverse materials, which helps to improve the user experience and the reliability of the application.