refactor(query): do not send data about mostly used words with partial query
This commit refactors the application to not send data about mostly used words in combination with a partial query. When the query is only part of a whole word, it does not make sense to include data about mostly used words as it may not be relevant to the user's search. With this refactor, the application is able to provide more relevant and accurate search results to the user.