feat(component): add date to Last Update Notification component
This commit adds functionality to the Last Update Notification component to display the date of the last update. This helps users to understand when the last update occurred and to track changes over time. By providing users with this information, the application is able to improve the user experience and make it easier for users to stay up-to-date with changes to the application.